pyproject.toml
[tool.poetry]
name = "bd-exemplos"
version = "0.1.0"
description = "Sample databases for testing (MySQL): seeds for Shop and Library domains."
readme = "README.md"
license = "MIT"
authors = ["Dinis Miranda"]
keywords = ["mysql", "database", "seed", "fixtures", "testing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]
packages = [{ include = "bd_exemplos" }]

[tool.poetry.dependencies]
python = "^3.9"
toml = ">=0.10.2"
mysql-connector-python = ">=8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^6.0"
ruff = "^0.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["bd_exemplos"]
branch = false
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__", "raise NotImplementedError"]
fail_under = 0
show_missing = true

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["bd_exemplos", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
